+++ R/inference.R
@@ -67,10 +67,15 @@ rater <- function(data,
   check_K(stan_data_list, model)
 
   # Create the full passed info for stan and the initialisation points.
-  stan_data <- c(stan_data_list, parse_priors(model, stan_data_list$K, method))
+  priors <- parse_priors(model, stan_data_list$K, stan_data_list$J)
+
+  if (method == "optim" && inherits(model, "dawid_skene")) {
+    check_beta_values(priors$beta)
+  }
+  stan_data <- c(stan_data_list, priors)
 
   if (is.null(inits)) {
-    inits <- creat_inits(model, stan_data_list)
+    inits <- create_inits(model, stan_data_list)
   }
 
   # TODO This could be made more complex if automatic switching is used.
@@ -139,44 +144,62 @@ as_stan_data <- function(data, data_format) {
 #'
 #' @param model The rater_model.
 #' @param K The number of categories in the data.
+#' @param J The number of raters in the data
 #' @param method The passed fitting method.
 #'
 #' @return The fully realised prior parameters
 #'
 #' @noRd
 #'
-parse_priors <- function(model, K, method) {
+parse_priors <- function(model, K, J) {
   switch(get_file(model),
-    "dawid_skene" = ds_parse_priors(model, K, method),
+    "dawid_skene" = ds_parse_priors(model, K, J),
     "class_conditional_dawid_skene" =
       class_conditional_ds_parse_priors(model, K),
     "hierarchical_dawid_skene" = hier_ds_parse_priors(model, K),
     stop("Unsupported model type", call. = FALSE))
 }
 
-ds_parse_priors <- function(model, K, method) {
+ds_parse_priors <- function(model, K, J) {
   pars <- get_parameters(model)
+
   # This is the default uniform prior taken from the Stan manual.
   if (is.null(pars$alpha)) {
     pars$alpha <- rep(3, K)
   }
+
+  # We need to alter the passed beta if:
+  # 1. It is a matrix - and we need to convert it into an array.
+  # 2. It is null - we need to create the default prior.
+  # Ideally this would be done earlier but we need to to know J. The matrix
+  # has already been validated i.e. it is square.
+
+  # 1.
+  # Convert from matrix to array.
+  if (is.matrix(pars$beta)) {
+    beta_slice <- pars$beta
+    pars$beta <- array(dim = c(J, K, K))
+    for (j in 1:J) {
+      pars$beta[j, , ] <- beta_slice
+    }
+  }
+
+  # 2.
   # This prior parameter is based on conjugate priors for the simplified model
-  # where the true class in known. Here we match on the mean.
+  # where the true class in known.
   if (is.null(pars$beta)) {
     N <- 8
     p <- 0.6
     on_diag <- N * p
     off_diag <- N * (1 - p) / (K - 1)
 
-    pars$beta <- matrix(off_diag, nrow = K, ncol = K)
-    diag(pars$beta) <- on_diag
-  }
+    beta_slice <- matrix(off_diag, nrow = K, ncol = K)
+    diag(beta_slice) <- on_diag
 
-  off_diag_problem <- any(pars$beta[row(pars$beta) != col(pars$beta)] <= 1.0)
-  if (method == "optim" && off_diag_problem) {
-      warning("Optimization may not converge if the off diagonal elements of ",
-              "beta are less than 1. Consider changing the prior parameters.",
-              call. = FALSE)
+    pars$beta <- array(dim = c(J,K,K))
+    for (j in 1:J) {
+      pars$beta[j, , ] <- beta_slice
+    }
   }
 
   pars
@@ -215,7 +238,7 @@ class_conditional_ds_parse_priors <- function(model, K) {
 #'
 #' @noRd
 #'
-creat_inits <- function(model, stan_data) {
+create_inits <- function(model, stan_data) {
   # better to have another short unique id...
   K <- stan_data$K
   J <- stan_data$J
@@ -452,3 +475,18 @@ validate_data <- function(data, data_format) {
   data
 }
 
+check_beta_values <- function(beta) {
+
+  J <- dim(beta)[[1]]
+  problems <- logical(J)
+  for (j in 1:J) {
+    beta_j <- beta[j, , ]
+    problems[[j]] <- any(beta_j[row(beta_j) != col(beta_j)] < 1.0)
+  }
+  off_diag_problem <- any(problems)
+  if (off_diag_problem) {
+      warning("Optimization may not converge if the off diagonal elements of ",
+              "beta are less than 1. Consider changing the prior parameters.",
+              call. = FALSE)
+  }
+}