+++ tests/testthat/test_rater.R
@@ -72,7 +72,8 @@ test_that("parse priors are correct", {
 
   # test default priors
   K <- anesthesia_list$K
-  ds_priors <- parse_priors(dawid_skene(), K, "mcmc")
+  J <- anesthesia_list$J
+  ds_priors <- parse_priors(dawid_skene(), K, J)
 
   # construct the default priors
   default_alpha <- rep(3, K)
@@ -81,32 +82,49 @@ test_that("parse priors are correct", {
   p <- 0.6
   on_diag <- N * p
   off_diag <- N * (1 - p) / (K - 1)
-  default_beta <- matrix(off_diag, nrow = K, ncol = K)
-  diag(default_beta) <- on_diag
+  beta_slice <- matrix(off_diag, nrow = K, ncol = K)
+  diag(beta_slice) <- on_diag
+  default_beta <- array(dim = c(J, K, K))
+  for (j in 1:J) {
+    default_beta[j, , ] <- beta_slice
+  }
 
   expect_equal(ds_priors$alpha, default_alpha)
   expect_equal(ds_priors$beta, default_beta)
 
   # test non-default priors
   test_alpha <- rep(9, K)
-  test_beta <- matrix(17, nrow = K, ncol = K)
-  test_beta_1 <- rep(1, K)
-  test_beta_2 <- rep(98, K)
-  ds_priors <- parse_priors(dawid_skene(alpha = test_alpha, beta = test_beta),
-                            K,
-                            "mcmc")
+  test_beta_mat <- matrix(17, nrow = K, ncol = K)
+  test_beta_array <- array(dim = c(J, K, K))
+  for (j in 1:J) {
+    test_beta_array[j, , ] <- test_beta_mat
+  }
 
-  expect_equal(ds_priors$alpha, test_alpha)
-  expect_equal(ds_priors$beta, test_beta)
+  ds_priors_mat <- parse_priors(dawid_skene(alpha = test_alpha,
+                                            beta = test_beta_mat),
+                                K,
+                                J)
+
+  expect_equal(ds_priors_mat$alpha, test_alpha)
+  expect_equal(ds_priors_mat$beta, test_beta_array)
+
+  ds_priors_array <- parse_priors(dawid_skene(alpha = test_alpha,
+                                              beta = test_beta_array),
+                                  K,
+                                  J)
+
+  expect_equal(ds_priors_array$beta, test_beta_array)
 
   # default priors
-  hds_priors <- parse_priors(hier_dawid_skene(), K, "mcmc")
+  hds_priors <- parse_priors(hier_dawid_skene(), K, J)
   expect_equal(hds_priors$alpha, default_alpha)
 
   # non-default priors
-  hds_priors <- parse_priors(hier_dawid_skene(alpha = test_alpha), K, "mcmc")
+  hds_priors <- parse_priors(hier_dawid_skene(alpha = test_alpha), K, J)
   expect_equal(hds_priors$alpha, test_alpha)
 
+  test_beta_1 <- rep(1, K)
+  test_beta_2 <- rep(98, K)
   # Class conditional model
   ccds_priors <- parse_priors(
     class_conditional_dawid_skene(
@@ -114,7 +132,7 @@ test_that("parse priors are correct", {
       beta_1 = test_beta_1,
       beta_2 = test_beta_2),
     K,
-    "mcmc"
+    J
   )
 
   expect_equal(ccds_priors$alpha, test_alpha)